feat  add close websocket
